+++ app/pb_kits/playbook/packs/examples.js
@@ -80,6 +80,7 @@ import * as TitleDetail from 'pb_title_detail/docs'
 import * as Toggle from 'pb_toggle/docs'
 import * as User from 'pb_user/docs'
 import * as UserBadge from 'pb_user_badge/docs'
+import * as WeekdayStacked from 'pb_weekday_stacked/docs'
 
 WebpackerReact.setup({
   ...Avatar,
@@ -149,4 +150,5 @@ WebpackerReact.setup({
   ...Toggle,
   ...User,
   ...UserBadge,
+  ...WeekdayStacked,
 })